Moves CheckApplicationName and CheckProjectName to cmd utils
