+++ pkg/odo/cli/application/describe.go
@@ -42,8 +42,8 @@ func (o *DescribeOptions) Complete(name string, cmd *cobra.Command, args []strin
 	if len(args) == 1 {
 		o.appName = args[0]
 	}
-	checkProjectName(o.Context)
-	checkApplicationName(o.appName)
+	util.CheckProjectName(o.Context.Project)
+	util.CheckApplicationName(o.appName)
 	return
 }
